+++ src/timeline_utils.py
@@ -1,10 +1,13 @@
-from datetime import datetime
+import datetime
+import random
 
 
-class TimelineUtils():
+class TimelineUtils:
     _MISS_VALUE_IN_TIMELINE_ERROR_MESSAGE = "Missing timeline value - ensure the timeline is in format '<video_number>,<start_time>,<end_time>' and separated by comma ','"
     _REDUNDANT_VALUE_IN_TIMELINE_ERROR_MESSAGE = "Redundant timeline value - ensure the timeline is in format '<video_number>,<start_time>,<end_time>' and separated by comma ','"
-    _START_TIME_AFTER_END_TIME_ERROR_MESSAGE = "Invalid timeline - start time is equal/after end time"
+    _START_TIME_AFTER_END_TIME_ERROR_MESSAGE = (
+        "Invalid timeline - start time is equal/after end time"
+    )
     _EMPTY_TIMELINE_ERROR_MESSAGE = "Invalid timeline - timeline cannot be empty"
 
     def __init__(self):
@@ -29,8 +32,10 @@ class TimelineUtils():
 
             for idx, parsed_timeline in enumerate(parsed_timeline_arr):
                 video, start, end = parsed_timeline
-                start_in_time = datetime.strptime(start, '%H:%M:%S').time()
-                end_in_time = datetime.strptime(end, '%H:%M:%S').time()
+                start_in_time = datetime.datetime.strptime(
+                    start, "%H:%M:%S").time()
+                end_in_time = datetime.datetime.strptime(
+                    end, "%H:%M:%S").time()
 
                 if start_in_time >= end_in_time:
                     return self._START_TIME_AFTER_END_TIME_ERROR_MESSAGE
@@ -43,3 +48,37 @@ class TimelineUtils():
                 return e
 
         return None
+
+    @staticmethod
+    def generate_random_segments(num_segments, min_per_segment, num_videos):
+        prev = None
+        start = "0:00:00"
+        end = 0
+        result = ""
+
+        for i in range(num_segments):
+            next_video_num = TimelineUtils.generate_next_video_num(
+                num_videos=num_videos, prev=prev)
+            end = str(datetime.timedelta(seconds=round(TimelineUtils.generate_next_segment_time(
+                min_per_segment=min_per_segment, prev_in_min=(i * min_per_segment)))))
+
+            result += f"{next_video_num},{start},{end}\n"
+
+            prev = next_video_num
+            start = end
+
+        return result
+
+    @staticmethod
+    def generate_next_video_num(num_videos, prev):
+        num = random.randint(1, num_videos)
+
+        if num == prev:
+            return TimelineUtils.generate_next_video_num(num_videos=num_videos, prev=prev)
+        else:
+            return num
+
+    @staticmethod
+    def generate_next_segment_time(min_per_segment, prev_in_min):
+        margin_in_sec = random.randint(-30, 30)
+        return (prev_in_min + min_per_segment) * 60 + margin_in_sec